Finish refactoring Context2D in to Canvas and remove Matrix
