--- src/dummy/toga_dummy/widgets/matrix.py
@@ -1,9 +1,0 @@
-from .base import Widget
-
-
-class Matrix(Widget):
-    def create(self):
-        pass
-
-    def transform_point(self, x, y):
-        pass
